+++ app/components/map/popups/MapPopupAirport.vue
@@ -64,7 +64,7 @@
             v-else
             class="airport_atc-popup"
             :controllers="getATC"
-            max-height="400px"
+            max-height="250px"
             :open-from="type === 'airport' ? 'bottom-center' : 'top-center'"
             :show-atis="type !== 'airport'"
             :show-facility="type === 'airport'"